+++ control/src/lib.rs
@@ -81,6 +81,7 @@ pub enum ControlAction {
     SetDelayQuantizationEight(bool),
     SetDelayHeadFeedbackAmount(usize, f32),
     SetDelayHeadVolume(usize, f32),
+    SetTone(f32),
 }
 
 #[derive(Clone, Copy, Debug)]
@@ -105,6 +106,7 @@ pub struct DSPAction {
     pub delay_head_position: [f32; 4],
     pub delay_head_feedback: [f32; 4],
     pub delay_head_volume: [f32; 4],
+    pub tone: f32,
 }
 
 impl From<DSPAction> for Attributes {
@@ -138,6 +140,7 @@ impl From<DSPAction> for Attributes {
             delay_head_2_volume: other.delay_head_volume[1],
             delay_head_3_volume: other.delay_head_volume[2],
             delay_head_4_volume: other.delay_head_volume[3],
+            tone: other.tone,
         }
     }
 }
@@ -149,6 +152,7 @@ pub struct Cache {
     pub hysteresis: HysteresisCache,
     pub wow: WowCache,
     pub delay: DelayCache,
+    pub tone: f32,
     pub leds: [ImpulseCache; 8],
     pub impulse: ImpulseCache,
 }
@@ -160,6 +164,7 @@ impl Default for Cache {
             hysteresis: HysteresisCache::default(),
             wow: WowCache::default(),
             delay: DelayCache::default(),
+            tone: 0.0,
             leds: [ImpulseCache::new(100); 8],
             impulse: ImpulseCache::new(100),
         }
@@ -277,6 +282,7 @@ pub fn cook_dsp_reaction_from_cache(cache: &Cache) -> DSPAction {
         ],
         delay_head_feedback: cache.delay.head_feedback,
         delay_head_volume: cache.delay.head_volume,
+        tone: cache.tone,
     }
 }
 
@@ -391,6 +397,9 @@ fn apply_control_action_in_cache(action: ControlAction, cache: &mut Cache) {
         }
         SetDelayHeadVolume(i, x) => {
             cache.delay.head_volume[i] = x;
+        }
+        SetTone(x) => {
+            cache.tone = x;
         }
     }
 }